Francisco branch

+++ src/main/java/com/example/qsproject/qsproject/repositories/UsersRespository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 // Isto faz com que o User tenha os métodos CRUD (create ,read ,update,delete)
-public interface UsersRespository extends JpaRepository<User,Integer> {
+public interface UsersRespository extends JpaRepository<User,Long> {
 }
